Rebuild native-lisp for each flavour.
It turns out native compilation always happens in $PWD/../native-lisp,
whereas elc files are produced in /lisp/ -- make only looks for the elc
files in /lisp/ and decides not to rebuild them again for each config.

This means that while we can have multiple build dirs, they both depend
on the srcdir in this regard, and native-lisp doesn't end up getting
re-built.